split point in time implementation from constructor
